+++ pom.xml
@@ -277,7 +277,7 @@
             <dependency>
                 <groupId>io.netty</groupId>
                 <artifactId>netty-bom</artifactId>
-                <version>4.2.4.Final</version>
+                <version>4.2.5.Final</version>
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
